Extract author parsing logic to new class
