+++ src/main/java/net/sf/jabref/model/entry/AuthorList.java
@@ -141,97 +141,18 @@ public class AuthorList {
     private final String[] authorsLastFirst = new String[4];
     private final String[] authorsLastFirstFirstLast = new String[2];
 
-    // The following variables are used only during parsing
-
-    private String original; // the raw bibtex author/editor field
-
-    // the following variables are updated by getToken procedure
-    private int tokenStart; // index in orig
-
-    private int tokenEnd; // to point 'abc' in ' abc xyz', start=2 and end=5
-
-    // the following variables are valid only if getToken returns TOKEN_WORD
-    private int tokenAbbr; // end of token abbreviation (always: token_start <
-
-    // token_abbr <= token_end)
-
-    private char tokenTerm; // either space or dash
-
-    private boolean tokenCase; // true if upper-case token, false if lower-case
-
-    // token
-
-    // Tokens of one author name.
-    // Each token occupies TGL consecutive entries in this vector (as described
-    // below)
-    private List<Object> tokens;
-
-    private static final int TOKEN_GROUP_LENGTH = 4; // number of entries for
-
-    // a token
-
-    // the following are offsets of an entry in a group of entries for one token
-    private static final int OFFSET_TOKEN = 0; // String -- token itself;
-
-    private static final int OFFSET_TOKEN_ABBR = 1; // String -- token abbreviation;
-
-    private static final int OFFSET_TOKEN_TERM = 2; // Character -- token terminator (either " " or
-    // "-") comma)
-
-    // Token types (returned by getToken procedure)
-    private static final int TOKEN_EOF = 0;
-
-    private static final int TOKEN_AND = 1;
-
-    private static final int TOKEN_COMMA = 2;
-
-    private static final int TOKEN_WORD = 3;
-
-    // Constant HashSet containing names of TeX special characters
-    private static final Set<String> TEX_NAMES = new HashSet<>();
-
-    // and static constructor to initialize it
-    static {
-        TEX_NAMES.add("aa");
-        TEX_NAMES.add("ae");
-        TEX_NAMES.add("l");
-        TEX_NAMES.add("o");
-        TEX_NAMES.add("oe");
-        TEX_NAMES.add("i");
-        TEX_NAMES.add("AA");
-        TEX_NAMES.add("AE");
-        TEX_NAMES.add("L");
-        TEX_NAMES.add("O");
-        TEX_NAMES.add("OE");
-        TEX_NAMES.add("j");
-    }
-
     private static final WeakHashMap<String, AuthorList> AUTHOR_CACHE = new WeakHashMap<>();
-
-
+    
     /**
-     * Parses the parameter strings and stores preformatted author information.
+     * Creates a new list of authors.
      * <p>
      * Don't call this constructor directly but rather use the getAuthorList()
      * method which caches its results.
      *
-     * @param bibtexAuthors contents of either <CODE>author</CODE> or <CODE>editor</CODE>
-     *                      bibtex field.
+     * @param authors the list of authors which should underlie this instance
      */
-    private AuthorList(String bibtexAuthors) {
-        authors = new ArrayList<>(5); // 5 seems to be reasonable initial size
-        original = bibtexAuthors; // initialization
-        tokenStart = 0;
-        tokenEnd = 0; // of parser
-        while (tokenStart < original.length()) {
-            Author author = getAuthor();
-            if (author != null) {
-                authors.add(author);
-            }
-        }
-        // clean-up
-        original = null;
-        tokens = null;
+    private AuthorList(List<Author> authors) {
+        this.authors = Objects.requireNonNull(authors);
     }
 
     /**
@@ -247,7 +168,8 @@ public class AuthorList {
 
         AuthorList authorList = AUTHOR_CACHE.get(authors);
         if (authorList == null) {
-            authorList = new AuthorList(authors);
+            AuthorListParser parser = new AuthorListParser();
+            authorList = new AuthorList(parser.parse(authors));
             AUTHOR_CACHE.put(authors, authorList);
         }
         return authorList;
@@ -323,303 +245,6 @@ public class AuthorList {
      */
     public static String fixAuthorNatbib(String authors) {
         return AuthorList.getAuthorList(authors).getAuthorsNatbib();
-    }
-
-    /**
-     * Parses one author name and returns preformatted information.
-     *
-     * @return Preformatted author name; <CODE>null</CODE> if author name is
-     * empty.
-     */
-    private Author getAuthor() {
-
-        tokens = new ArrayList<>(); // initialization
-        int vonStart = -1;
-        int lastStart = -1;
-        int commaFirst = -1;
-        int commaSecond = -1;
-
-        // First step: collect tokens in 'tokens' Vector and calculate indices
-        boolean continueLoop = true;
-        while (continueLoop) {
-            int token = getToken();
-            switch (token) {
-            case TOKEN_EOF:
-            case TOKEN_AND:
-                continueLoop = false;
-                break;
-            case TOKEN_COMMA:
-                if (commaFirst < 0) {
-                    commaFirst = tokens.size();
-                } else if (commaSecond < 0) {
-                    commaSecond = tokens.size();
-                }
-                break;
-            case TOKEN_WORD:
-                tokens.add(original.substring(tokenStart, tokenEnd));
-                tokens.add(original.substring(tokenStart, tokenAbbr));
-                tokens.add(tokenTerm);
-                tokens.add(tokenCase);
-                if (commaFirst >= 0) {
-                    break;
-                }
-                if (lastStart >= 0) {
-                    break;
-                }
-                if (vonStart < 0) {
-                    if (!tokenCase) {
-                        vonStart = tokens.size() - AuthorList.TOKEN_GROUP_LENGTH;
-                        break;
-                    }
-                } else if ((lastStart < 0) && tokenCase) {
-                    lastStart = tokens.size() - AuthorList.TOKEN_GROUP_LENGTH;
-                    break;
-                }
-                break;
-            default:
-                break;
-            }
-        }
-
-        // Second step: split name into parts (here: calculate indices
-        // of parts in 'tokens' Vector)
-        if (tokens.isEmpty()) {
-            return null; // no author information
-        }
-
-        // the following negatives indicate absence of the corresponding part
-        int firstPartStart = -1;
-        int vonPartStart = -1;
-        int lastPartStart = -1;
-        int jrPartStart = -1;
-        int firstPartEnd;
-        int vonPartEnd = 0;
-        int lastPartEnd = 0;
-        int jrPartEnd = 0;
-        if (commaFirst < 0) { // no commas
-            if (vonStart < 0) { // no 'von part'
-                lastPartEnd = tokens.size();
-                lastPartStart = tokens.size() - AuthorList.TOKEN_GROUP_LENGTH;
-                int index = (tokens.size() - (2 * AuthorList.TOKEN_GROUP_LENGTH)) + AuthorList.OFFSET_TOKEN_TERM;
-                if (index > 0) {
-                    Character ch = (Character) tokens.get(index);
-                    if (ch == '-') {
-                        lastPartStart -= AuthorList.TOKEN_GROUP_LENGTH;
-                    }
-                }
-                firstPartEnd = lastPartStart;
-                if (firstPartEnd > 0) {
-                    firstPartStart = 0;
-                }
-            } else { // 'von part' is present
-                if (lastStart >= 0) {
-                    lastPartEnd = tokens.size();
-                    lastPartStart = lastStart;
-                    vonPartEnd = lastPartStart;
-                } else {
-                    vonPartEnd = tokens.size();
-                }
-                vonPartStart = vonStart;
-                firstPartEnd = vonPartStart;
-                if (firstPartEnd > 0) {
-                    firstPartStart = 0;
-                }
-            }
-        } else { // commas are present: it affects only 'first part' and
-            // 'junior part'
-            firstPartEnd = tokens.size();
-            if (commaSecond < 0) { // one comma
-                if (commaFirst < firstPartEnd) {
-                    firstPartStart = commaFirst;
-                }
-            } else { // two or more commas
-                if (commaSecond < firstPartEnd) {
-                    firstPartStart = commaSecond;
-                }
-                jrPartEnd = commaSecond;
-                if (commaFirst < jrPartEnd) {
-                    jrPartStart = commaFirst;
-                }
-            }
-            if (vonStart == 0) { // 'von part' is present
-                if (lastStart < 0) {
-                    vonPartEnd = commaFirst;
-                } else {
-                    lastPartEnd = commaFirst;
-                    lastPartStart = lastStart;
-                    vonPartEnd = lastPartStart;
-                }
-                vonPartStart = 0;
-            } else { // no 'von part'
-                lastPartEnd = commaFirst;
-                if (lastPartEnd > 0) {
-                    lastPartStart = 0;
-                }
-            }
-        }
-
-        if ((firstPartStart == -1) && (lastPartStart == -1) && (vonPartStart != -1)) {
-            // There is no first or last name, but we have a von part. This is likely
-            // to indicate a single-entry name without an initial capital letter, such
-            // as "unknown".
-            // We make the von part the last name, to facilitate handling by last-name formatters:
-            lastPartStart = vonPartStart;
-            lastPartEnd = vonPartEnd;
-            vonPartStart = -1;
-            vonPartEnd = -1;
-        }
-
-        // Third step: do actual splitting, construct Author object
-        return new Author(firstPartStart < 0 ? null : concatTokens(firstPartStart,
-                firstPartEnd, AuthorList.OFFSET_TOKEN, false), firstPartStart < 0 ? null : concatTokens(
-                firstPartStart, firstPartEnd, AuthorList.OFFSET_TOKEN_ABBR, true), vonPartStart < 0 ? null
-                : concatTokens(vonPartStart, vonPartEnd, AuthorList.OFFSET_TOKEN, false),
-                lastPartStart < 0 ? null : concatTokens(lastPartStart, lastPartEnd,
-                        AuthorList.OFFSET_TOKEN, false), jrPartStart < 0 ? null : concatTokens(jrPartStart,
-                jrPartEnd, AuthorList.OFFSET_TOKEN, false));
-    }
-
-    /**
-     * Concatenates list of tokens from 'tokens' Vector. Tokens are separated by
-     * spaces or dashes, depending on stored in 'tokens'. Callers always ensure
-     * that start < end; thus, there exists at least one token to be
-     * concatenated.
-     *
-     * @param start     index of the first token to be concatenated in 'tokens' Vector
-     *                  (always divisible by TOKEN_GROUP_LENGTH).
-     * @param end       index of the first token not to be concatenated in 'tokens'
-     *                  Vector (always divisible by TOKEN_GROUP_LENGTH).
-     * @param offset    offset within token group (used to request concatenation of
-     *                  either full tokens or abbreviation).
-     * @param dotAfter <CODE>true</CODE> -- add period after each token, <CODE>false</CODE> --
-     *                  do not add.
-     * @return the result of concatenation.
-     */
-    private String concatTokens(int start, int end, int offset, boolean dotAfter) {
-        StringBuilder result = new StringBuilder();
-        // Here we always have start < end
-        result.append((String) tokens.get(start + offset));
-        if (dotAfter) {
-            result.append('.');
-        }
-        int updatedStart = start + AuthorList.TOKEN_GROUP_LENGTH;
-        while (updatedStart < end) {
-            result.append(tokens.get((updatedStart - AuthorList.TOKEN_GROUP_LENGTH) + AuthorList.OFFSET_TOKEN_TERM));
-            result.append((String) tokens.get(updatedStart + offset));
-            if (dotAfter) {
-                result.append('.');
-            }
-            updatedStart += AuthorList.TOKEN_GROUP_LENGTH;
-        }
-        return result.toString();
-    }
-
-    /**
-     * Parses the next token.
-     * <p>
-     * The string being parsed is stored in global variable <CODE>orig</CODE>,
-     * and position which parsing has to start from is stored in global variable
-     * <CODE>token_end</CODE>; thus, <CODE>token_end</CODE> has to be set
-     * to 0 before the first invocation. Procedure updates <CODE>token_end</CODE>;
-     * thus, subsequent invocations do not require any additional variable
-     * settings.
-     * <p>
-     * The type of the token is returned; if it is <CODE>TOKEN_WORD</CODE>,
-     * additional information is given in global variables <CODE>token_start</CODE>,
-     * <CODE>token_end</CODE>, <CODE>token_abbr</CODE>, <CODE>token_term</CODE>,
-     * and <CODE>token_case</CODE>; namely: <CODE>orig.substring(token_start,token_end)</CODE>
-     * is the thext of the token, <CODE>orig.substring(token_start,token_abbr)</CODE>
-     * is the token abbreviation, <CODE>token_term</CODE> contains token
-     * terminator (space or dash), and <CODE>token_case</CODE> is <CODE>true</CODE>,
-     * if token is upper-case and <CODE>false</CODE> if token is lower-case.
-     *
-     * @return <CODE>TOKEN_EOF</CODE> -- no more tokens, <CODE>TOKEN_COMMA</CODE> --
-     * token is comma, <CODE>TOKEN_AND</CODE> -- token is the word
-     * "and" (or "And", or "aND", etc.), <CODE>TOKEN_WORD</CODE> --
-     * token is a word; additional information is given in global
-     * variables <CODE>token_start</CODE>, <CODE>token_end</CODE>,
-     * <CODE>token_abbr</CODE>, <CODE>token_term</CODE>, and
-     * <CODE>token_case</CODE>.
-     */
-    private int getToken() {
-        tokenStart = tokenEnd;
-        while (tokenStart < original.length()) {
-            char c = original.charAt(tokenStart);
-            if (!((c == '~') || (c == '-') || Character.isWhitespace(c))) {
-                break;
-            }
-            tokenStart++;
-        }
-        tokenEnd = tokenStart;
-        if (tokenStart >= original.length()) {
-            return AuthorList.TOKEN_EOF;
-        }
-        if (original.charAt(tokenStart) == ',') {
-            tokenEnd++;
-            return AuthorList.TOKEN_COMMA;
-        }
-        tokenAbbr = -1;
-        tokenTerm = ' ';
-        tokenCase = true;
-        int bracesLevel = 0;
-        int currentBackslash = -1;
-        boolean firstLetterIsFound = false;
-        while (tokenEnd < original.length()) {
-            char c = original.charAt(tokenEnd);
-            if (c == '{') {
-                bracesLevel++;
-            }
-            if ((c == '}') && (bracesLevel > 0)) {
-                bracesLevel--;
-            }
-            if (firstLetterIsFound && (tokenAbbr < 0) && (bracesLevel == 0)) {
-                tokenAbbr = tokenEnd;
-            }
-            if (!firstLetterIsFound && (currentBackslash < 0) && Character.isLetter(c)) {
-                if (bracesLevel == 0) {
-                    tokenCase = Character.isUpperCase(c);
-                } else {
-                    // If this is a particle in braces, always treat it as if it starts with
-                    // an upper case letter. Otherwise a name such as "{van den Bergen}, Hans"
-                    // will not yield a proper last name:
-                    tokenCase = true;
-                }
-                firstLetterIsFound = true;
-            }
-            if ((currentBackslash >= 0) && !Character.isLetter(c)) {
-                if (!firstLetterIsFound) {
-                    String texCmdName = original.substring(currentBackslash + 1, tokenEnd);
-                    if (AuthorList.TEX_NAMES.contains(texCmdName)) {
-                        tokenCase = Character.isUpperCase(texCmdName.charAt(0));
-                        firstLetterIsFound = true;
-                    }
-                }
-                currentBackslash = -1;
-            }
-            if (c == '\\') {
-                currentBackslash = tokenEnd;
-            }
-            if ((bracesLevel == 0) && ((",~-".indexOf(c) != -1) || Character.isWhitespace(c))) {
-                break;
-            }
-            // Morten Alver 18 Apr 2006: Removed check for hyphen '-' above to
-            // prevent
-            // problems with names like Bailey-Jones getting broken up and
-            // sorted wrong.
-            // Aaron Chen 14 Sep 2008: Enable hyphen check for first names like Chang-Chin
-            tokenEnd++;
-        }
-        if (tokenAbbr < 0) {
-            tokenAbbr = tokenEnd;
-        }
-        if ((tokenEnd < original.length()) && (original.charAt(tokenEnd) == '-')) {
-            tokenTerm = '-';
-        }
-        if ("and".equalsIgnoreCase(original.substring(tokenStart, tokenEnd))) {
-            return AuthorList.TOKEN_AND;
-        } else {
-            return AuthorList.TOKEN_WORD;
-        }
     }
 
     /**
